pyproject.toml
[project]
name = "nonebot-plugin-qqgroup-filedownloader"
version = "0.1.0"
description = "✨ 一个基于 NoneBot2 的插件，用于监听并下载QQ群上传的文件 ✨"
readme = "README.md"
requires-python = ">=3.9, <4.0"
dependencies = ["nonebot-adapter-onebot>=2.4.6", "nonebot2[fastapi]>=2.4.3", "aiohttp>=3.12.15", "nonebot-plugin-localstore>=0.7.4"]
license = {text = "MIT"}

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"


[tool.pdm]
distribution = true
